Updated codebase to compile under Visual Studio again.
Added pthread and readline libraries for Windows.
Cleaned up the third-party library referencing system.
Changed SDP functionality in the debugger to be activated only when -DEXPERIMENTAL=1 is passed to CMake.
Renamed aux.c / aux.h to dbgaux.c / dbgaux.h to prevent conflicts on Windows.

+++ libdcpu-ci-lang-c/nodes/NInteger.cpp
@@ -47,7 +47,7 @@ AsmBlock* NInteger::compile(AsmGenerator& context)
 	AsmBlock* block = new AsmBlock();
 
 	// Load the numeric value in register A.
-	*block <<	"	SET A, " << this->value << std::endl;
+	*block <<	"	SET A, " << (const size_t)this->value << std::endl;
 
 	return block;
 }
